Edited Tasks and Events
When tasks/events are displayed, they are now filtered by the user id of the person who uploaded them.
Similarly when tasks/events are added, the id of the current user is stored in the database as well.
Also made minor changes to actions after a form is submitted, e.g. more flash messages